pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "quran-persian-chatbot"
version = "0.1.0"
description = "Persian Quran QA chatbot using RAG, ChromaDB and OpenRouter"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Mahdi" }
]
dependencies = [
  "chromadb>=0.5.0",
  "llama-index-core>=0.12.0",
  "llama-index-embeddings-huggingface>=0.5.0",
  "llama-index-readers-file>=0.4.0",
  "llama-index-vector-stores-chroma>=0.4.0",
  "openai>=1.52.0",
  "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.6.0",
]

[project.scripts]
quran-chatbot = "quran_persian_chatbot.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
